requirements.txt
streamlit
google-generativeai
youtube-transcript-api
python-dotenv
pytube
toml